rename vnni.py to x86.py
