+++ tests/python/unittest/test_tir_schedule_tensorize.py
@@ -22,7 +22,7 @@ import tvm.testing
 from tvm import tir, te
 from tvm.script import tir as T
 from tvm.tir.schedule.testing import verify_trace_roundtrip
-from tvm.tir.tensor_intrin.vnni import INTRIN_NAME as VNNI_INTRIN
+from tvm.tir.tensor_intrin.x86 import VNNI_INTRIN
 
 # fmt: off
 # pylint: disable=no-member,invalid-name,unused-variable,line-too-long,redefined-outer-name,unexpected-keyword-arg,too-many-nested-blocks
